pyproject.toml
[tool.poetry]
name = "pfhedge"
version = "0.20.0"
description = "Deep Hedging in PyTorch"
authors = ["Shota Imaki <shota.imaki.0801@gmail.com>"]
license = "MIT"
repository = "https://github.com/pfnet-research/pfhedge"

[tool.poetry.dependencies]
python = "^3.8.0"
torch = "^1.9.0"
tqdm = "^4.62.3"
qiskit = {extras = ["machine-learning"], version = "^0.43.0"}
qiskit-aqt-provider = "^0.4.3"
seaborn = "^0.12.2"
pennylane-qiskit = "^0.29.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^22.3.0"
isort = "^5.9.3"
mypy = "^0.910"
pytest-cov = "^3.0.0"
Sphinx = "^4.2.0"
sphinx-autobuild = "^2021.3.14"
sphinx-copybutton = "^0.4.0"
furo = "^2021.9.22"
codecov = "^2.1.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
skip-magic-trailing-comma = true

[tool.isort]
profile = 'black'
force_single_line = true

[tool.pytest.ini_options]
addopts = "-m \"not gpu\""
markers = "gpu: Tests that require GPU"
